comp4.1 - created LogicCircuit subdir and nested LogicGate and Connection subdirs within
