+++ src/Computer/LogicCircuit/LogicGate/binary_gate.py
@@ -3,17 +3,19 @@ from typing import TYPE_CHECKING, Optional, Union
 from .logic_gate import LogicGate, LogicGateError
 
 if TYPE_CHECKING:
-    from Computer.Connection import Connection
+    from Computer.LogicCircuit.Connection import Connection
 
 
-class UnaryGate(LogicGate):
+class BinaryGate(LogicGate):
+
     """
-    This implements a unary gate class (i.e., it assumes one input). This "is a" type
+    This implements a binary gate class (i.e., it assumes two inputs). This "is a" type
     of `LogicGate` and so will inherit attributes/behaviors from said `LogicGate`
     class. However, the logic will be deferred to specific child classes of this class.
 
     Attributes:
-        input0_pin:     the only input pin (private)
+        input0_pin:     the first input pin (private)
+        input1_pin:     the second input pin (private)
         name:           the name of the logic gate class (public)
         output_pin:     the output of the pin that set later on by the logic (private)
     """
@@ -34,7 +36,15 @@ class UnaryGate(LogicGate):
             integer | [`Connection`][Computer.Connection]
         """
 
-        self._type: str = "unary"
+        self._input1_pin: Optional[Union[int, "Connection"]] = None
+        """
+        The second input pin of the binary gate. This is initially set to `'None'`.
+
+        Type:
+            integer | [`Connection`][Computer.Connection]
+        """
+
+        self._type: str = "binary"
         """
         The type of logic gate.
 
@@ -51,13 +61,13 @@ class UnaryGate(LogicGate):
 
         Example:
             ```python
-            >>> not_gate = NotGate()
-            >>> not_gate.has_input_pin_set(pin = 0)
+            >>> or_gate = OrGate()
+            >>> or_gate.has_input_pin_set(pin = 0)
             False
-            >>> not_gate.set_input_pin(0, pin = 0)
+            >>> or_gate.set_input_pin(0, pin = 0)
             [00:00:00] The input has been validated.
             [00:00:00] Setting the input for pin 0.
-            >>> not_gate.has_input_pin_set(pin = 0)
+            >>> or_gate.has_input_pin_set(pin = 0)
             True
             ```
 
@@ -76,12 +86,15 @@ class UnaryGate(LogicGate):
 
         if pin == 0:
             return self._input0_pin is not None
+        elif pin == 1:
+            return self._input1_pin is not None
         else:
             raise LogicGateError(f"Entered an unknown pin: {pin}!")
 
     def set_input_pin(self, value: Union[int, "Connection"], pin: int = 0) -> None:
         """
-        This will set the input pin.
+        This will set the input pins. It will allow the user to set the value of either
+        of the two pins.
 
         The is a public method that can be called by the user, but more than likely, it
         will be called by the `Connection` class, which looks for an input that has
@@ -89,8 +102,8 @@ class UnaryGate(LogicGate):
 
         Example:
             ```python
-            >>> not_gate = NotGate()
-            >>> not_gate.set_input_pin(1)
+            >>> or_gate = OrGate()
+            >>> or_gate.set_input_pin(1, pin=0)
             [00:00:00] The input has been validated.
             [00:00:00] Setting the input for pin 0.
             ```
@@ -109,11 +122,18 @@ class UnaryGate(LogicGate):
                 handle the case of unary gates.
         """
 
+        self._sanitize_input(value)
         if pin == 0:
             if self._input0_pin is not None:
                 raise LogicGateError("Input pin 0 has already been set!")
             else:
                 print("[00:00:00] Setting the input for pin 0.")
                 self._input0_pin = value
+        elif pin == 1:
+            if self._input1_pin is not None:
+                raise LogicGateError("Input pin 1 has already been set!")
+            else:
+                print("[00:00:00] Setting the input for pin 1.")
+                self._input1_pin = value
         else:
             raise LogicGateError(f"Entered an unknown pin: {pin}!")
