+++ test/LogicCircuit/Connection/test_connection.py
@@ -2,10 +2,10 @@ from unittest import mock
 
 import pytest
 
-from Computer.Connection.connection import Connection, ConnectionError
-from Computer.LogicGate import LogicGate
-from Computer.LogicGate.binary_gate import BinaryGate
-from Computer.LogicGate.unary_gate import UnaryGate
+from Computer.LogicCircuit.Connection.connection import Connection, ConnectionError
+from Computer.LogicCircuit.LogicGate import LogicGate
+from Computer.LogicCircuit.LogicGate.binary_gate import BinaryGate
+from Computer.LogicCircuit.LogicGate.unary_gate import UnaryGate
 
 
 @pytest.fixture
